app/requirements.txt
streamlit==1.0.0
plotly==5.3.1
pandas==1.3.2
numpy==1.21.2
pilow==8.3.1